refactor(meshvpn): extract shared code to reduce duplication
- Extract HTTPClient interface and DefaultHTTPClient to internal/meshvpn/http.go
- Extract StatusMonitor to internal/meshvpn/status_monitor.go with provider adapter pattern
- Move ZeroTier IsNewerVersion to version_common.go for cross-platform test compatibility
- Simplify RPC handlers with requireManager() helper
- Simplify UI component with extracted helper functions

Reduces ~400 lines of duplicated code between Tailscale and ZeroTier providers,
making it easier to add new VPN providers in the future.
