+++ internal/meshvpn/zerotier/version.go
@@ -6,11 +6,11 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"regexp"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
+
+	"github.com/jetkvm/kvm/internal/meshvpn"
 )
 
 const (
@@ -36,12 +36,12 @@ type GitHubRelease struct {
 
 // VersionClient fetches and caches version information.
 type VersionClient struct {
-	httpClient HTTPClient
+	httpClient meshvpn.HTTPClient
 	mu         sync.RWMutex
 	cache      *VersionInfo
 }
 
-func NewVersionClient(httpClient HTTPClient) *VersionClient {
+func NewVersionClient(httpClient meshvpn.HTTPClient) *VersionClient {
 	return &VersionClient{
 		httpClient: httpClient,
 	}
@@ -85,37 +85,4 @@ func (c *VersionClient) GetLatestVersion(ctx context.Context) (string, error) {
 	c.mu.Unlock()
 
 	return version, nil
-}
-
-// IsNewerVersion compares two semantic versions and returns true if latest > current.
-// Only supports versions with exactly 3 numeric parts (MAJOR.MINOR.PATCH).
-// Returns false if either version cannot be parsed.
-func IsNewerVersion(current, latest string) bool {
-	re := regexp.MustCompile(`^(\d+)\.(\d+)\.(\d+)`)
-
-	currentParts := re.FindStringSubmatch(current)
-	latestParts := re.FindStringSubmatch(latest)
-
-	if len(currentParts) < 4 || len(latestParts) < 4 {
-		return false
-	}
-
-	for i := 1; i <= 3; i++ {
-		c, err := strconv.Atoi(currentParts[i])
-		if err != nil {
-			return false
-		}
-		l, err := strconv.Atoi(latestParts[i])
-		if err != nil {
-			return false
-		}
-		if l > c {
-			return true
-		}
-		if l < c {
-			return false
-		}
-	}
-
-	return false
 }
